# zadanie 4

+++ app/models/Payment.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json.Json
 
-case class Payment(id: Int, value: Float, paid : Boolean)
+case class Payment(id: Int, value: Int, paid : Boolean)
 
 object Payment {
   implicit val paymentFormat = Json.format[Payment]
